requirements.txt
Django==4.1
gunicorn==20.0.4
django-heroku==0.3.1
redis==4.3.4
django-redis==5.2.0
psycopg2==2.9.3
python-decouple==3.6
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.1
dj-rest-auth==2.1.7
django-cors-headers==3.5.0
drf-spectacular==0.13.2
whitenoise==5.2.0
django-allauth==0.43.0
yfinance==0.1.74
yahoo_fin-0.8.9.1